Split Log Types
